% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adlaw.R
\docType{package}
\name{adlaw}
\alias{adlaw-package}
\alias{adlaw}
\title{An Interface to the Scaling Up Nutrition (SUN) Movement Datasets}
\description{
Scaling Up Nutrition (SUN) \url{http://scalingupnutrition.org} is a
global movement of 60 countries aiming to end malnutrition in all its forms.
The SUN Movement. This package contains various datasets produced by the
Movement and functions that aid in the manipulation and processing of these
datasets. It also includes functions that support analysis and visualisation
of these datasets. Finally, a Shiny application is included to allow user to
interactively engage with the publicly-available SUN Movement datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katilingban/adlaw}
  \item \url{https://katilingban.io/scalingupnutrition/}
  \item \url{https://katilingban.io/adlaw/}
  \item Report bugs at \url{https://github.com/katilingban/adlaw/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
