% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{investment}
\alias{investment}
\title{Scaling Up Nutrition investment database}
\format{
A tibble with 14 columns and 16313 rows.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{Country} \tab Country name \cr
   \emph{Budget year} \tab Year (as character value) budget is for \cr
   \emph{Ministry, department or agency} \tab Name of ministry, department or agency \cr
   \emph{Search term} \tab Search term \cr
   \emph{Budget line item} \tab Name of budget line item \cr
   \emph{Budget line item description} \tab Description of budget line item \cr
   \emph{Nutrition-Sensitive/Specific} \tab Classification of budget line item \cr
   \emph{Sector} \tab Sector classification for budget line item \cr
   \emph{Nutrition-sensitive typology} \tab Type of nutrition-sensitive budget line item \cr
   \emph{Nutrition-specific typology} \tab Type of nutrition-specific budget line item \cr
   \emph{Currency code} \tab Thee letter code for country currency \cr
   \emph{Allocations (Country currency)} \tab Amount in country currency (as character value) allocated to budget line item \cr
   \emph{Deflator allocations (Country currency)} \tab Amount in country currency (as numeric value) allocated to budget line item \cr
   \emph{Available sub-national information} \tab Is there sub-national information available? \cr
}
}
\usage{
investment
}
\description{
Scaling Up Nutrition investment database
}
\keyword{datasets}
