% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_report_2015}
\alias{progress_report_2015}
\title{progress_report_2015}
\format{A dataframe with 4 columns and 10886 rows:
\describe{
\item{\code{text}}{Line of text from the 2015 annual progress report}
\item{\code{page}}{Page number to which line of text can be found}
\item{\code{linenumber}}{Line number of text}
\item{\code{chapter}}{Document chapter to which line of text can be found}
}}
\source{
Scaling Up Nutrition Movement 2015
}
\usage{
progress_report_2015
}
\description{
SUN movement annual progress report 2015
}
\keyword{datasets}
