% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalingupnutrition.R
\docType{package}
\name{scalingupnutrition}
\alias{scalingupnutrition}
\alias{scalingupnutrition-package}
\title{Scaling Up Nutrition (SUN) Movement Annual Progress Reports}
\description{
Scaling Up Nutrition (SUN) \url{http://scalingupnutrition.org} is a global movement
of 60 countries aiming to end malnutrition in all its forms. The SUN Movement
Annual Progress Report provides a detailed snapshot of the progress and results
towards scaling up nutrition in SUN Countries and the broad evolution of the
Movement. This package contains text dataset of the Movement's annual reports
from 2011-2017 and various functions for processing, manipulating and analysing
this dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katilingban/scalingupnutrition}
  \item Report bugs at \url{https://github.com/katilingban/scalingupnutrition/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
