% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_report_2014}
\alias{progress_report_2014}
\title{SUN movement annual progress report 2014}
\format{
A dataframe with 4 columns and 3873 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{text} \tab Line of text from the 2014 annual progress report \cr
   \emph{page} \tab Page number to which line of text can be found \cr
   \emph{linenumber} \tab Line number of text \cr
   \emph{chapter} \tab Document chapter to which line of text can be found \cr
}
}
\source{
Scaling Up Nutrition Movement 2014
}
\usage{
progress_report_2014
}
\description{
SUN movement annual progress report 2014
}
\keyword{datasets}
