% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_report_2016}
\alias{progress_report_2016}
\title{SUN Movement annual progress report 2016}
\format{
A tibble with 4 columns and 8070 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{text} \tab Line of text from the 2016 annual progress report \cr
   \emph{page} \tab Page number to which line of text can be found \cr
   \emph{linenumber} \tab Line number of text \cr
   \emph{chapter} \tab Document chapter to which line of text can be found \cr
}
}
\source{
Scaling Up Nutrition Movement 2016
}
\usage{
progress_report_2016
}
\description{
SUN Movement annual progress report 2016
}
\examples{
progress_report_2016

}
\keyword{datasets}
