% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{progress_report_2017}
\alias{progress_report_2017}
\title{SUN Movement annual progress report 2017}
\format{
A tibble with 4 columns and 7586 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{text} \tab Line of text from the 2017 annual progress report \cr
   \emph{page} \tab Page number to which line of text can be found \cr
   \emph{linenumber} \tab Line number of text \cr
   \emph{chapter} \tab Document chapter to which line of text can be found \cr
}
}
\source{
Scaling Up Nutrition Movement 2017
}
\usage{
progress_report_2017
}
\description{
SUN Movement annual progress report 2017
}
\examples{
progress_report_2017

}
\keyword{datasets}
